requirements.txt
requests
boto3
python-ulid